Conflicts:
	src/server/scripts/Northrend/Ulduar/ulduar/boss_ignis.cpp

+++ sql/fc_updates/anticheat_characters.sql
@@ -3,7 +3,7 @@ DROP TABLE IF EXISTS `players_reports_status`;
 CREATE TABLE `players_reports_status` (
   `guid` int(10) unsigned NOT NULL DEFAULT '0',
   `creation_time` int(10) unsigned NOT NULL DEFAULT '0',
-  `average` bigint(20) unsigned NOT NULL DEFAULT '0',
+  `average` float NOT NULL DEFAULT '0',
   `total_reports` bigint(20) unsigned NOT NULL DEFAULT '0',
   `speed_reports` bigint(20) unsigned NOT NULL DEFAULT '0',
   `fly_reports` bigint(20) unsigned NOT NULL DEFAULT '0',
